Temporary fix to make type checker happy

+++ thrift/types.thrift
@@ -1,5 +1,5 @@
 struct Task {
   1: required string desc,
   2: required i16 pri, // 0 being highest priority
-  3: required i64 time, // UNIX timestamp. 0 to indicate unset
+  3: required double time, // UNIX timestamp. 0 to indicate unset
 }
